重构预览部分
